# src/main/java/com/g5/teamtaskmanagement/dto/request/ChangePasswordRequest.java
package com.g5.teamtaskmanagement.dto.request;

import jakarta.validation.constraints.AssertTrue;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Pattern;

public class ChangePasswordRequest {
    @NotBlank
    private String oldPassword;

    @NotBlank
    @Pattern(regexp = "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[^A-Za-z0-9]).{8,}$", message = "must contain at least 8 characters, including uppercase, lowercase, number, and special character")
    @NotBlank
    private String newPassword;

    @NotBlank
    @Pattern(regexp = "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[^A-Za-z0-9]).{8,}$", message = "must contain at least 8 characters, including uppercase, lowercase, number, and special character")
    private String rePassword;

    public String getOldPassword() {
        return oldPassword;
    }

    public void setOldPassword(String oldPassword) {
        this.oldPassword = oldPassword;
    }

    public String getNewPassword() {
        return newPassword;
    }

    public void setNewPassword(String newPassword) {
        this.newPassword = newPassword;
    }

    public String getRePassword() {
        return rePassword;
    }

    public void setRePassword(String rePassword) {
        this.rePassword = rePassword;
    }

    @AssertTrue(message = "newPassword and rePassword must match")
    public boolean isPasswordConfirmationValid() {
        return newPassword != null && newPassword.equals(rePassword);
    }
}