# src/main/java/com/g5/teamtaskmanagement/dto/request/RefreshRequest.java
package com.g5.teamtaskmanagement.dto.request;

import jakarta.validation.constraints.NotBlank;

public class RefreshRequest {
    @NotBlank
    private String refreshToken;

    public String getRefreshToken() { return refreshToken; }
    public void setRefreshToken(String refreshToken) { this.refreshToken = refreshToken; }
}