# team-tast-management/api-docs.json
{"openapi":"3.1.0","info":{"title":"Team Task Management API","description":"API documentation for the Team Task Management project","contact":{"name":"G5 Team"},"license":{"name":"Apache 2.0"},"version":"v1"},"externalDocs":{"description":"Project source","url":"https://example.com"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/users/me":{"get":{"tags":["user-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserDto"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserDto"}}}}}}},"/api/users/me/password":{"put":{"tags":["user-controller"],"operationId":"changePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseObject"}}}}}}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseObject"}}}}}}},"/api/auth/refresh":{"post":{"tags":["auth-controller"],"operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseObject"}}}}}}},"/api/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseObject"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponse"}}}}}}}},"components":{"schemas":{"UpdateProfileRequest":{"type":"object","properties":{"fullName":{"type":"string","minLength":1},"avatarUrl":{"type":"string"}},"required":["fullName"]},"ApiResponseUserDto":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/UserDto"},"errors":{"type":"array","items":{"type":"string"}}}},"UserDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"fullName":{"type":"string"}}},"ChangePasswordRequest":{"type":"object","properties":{"oldPassword":{"type":"string","minLength":1},"newPassword":{"type":"string","minLength":1}},"required":["newPassword","oldPassword"]},"ApiResponseObject":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{},"errors":{"type":"array","items":{"type":"string"}}}},"RegisterRequest":{"type":"object","properties":{"email":{"type":"string","minLength":1},"password":{"type":"string","minLength":1},"fullName":{"type":"string","minLength":1}},"required":["email","fullName","password"]},"RefreshRequest":{"type":"object","properties":{"refreshToken":{"type":"string","minLength":1}},"required":["refreshToken"]},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"ApiResponseAuthResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/AuthResponse"},"errors":{"type":"array","items":{"type":"string"}}}},"AuthResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"user":{"$ref":"#/components/schemas/UserDto"}}}}}}